refactor: deepen editor architecture
